+++ CHANGELOG.md
@@ -37,6 +37,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 * [#93](https://github.com/babylonlabs-io/btc-staker/pull/93) Fix linting config
 * [#95](https://github.com/babylonlabs-io/btc-staker/pull/95) Bump babylon to v0.16
+* [#96](https://github.com/babylonlabs-io/btc-staker/pull/96) Check allow list
+expiration height before creating new delegations
 
 ## v0.10.0